refactor(Routes)

+++ src/Routes/Products/Categories/routes.ts
@@ -4,14 +4,15 @@ import { GetProductsCategoriesController } from "../../../Controllers/Products/C
 import { DeleteProductCategoryController } from "../../../Controllers/Products/Categories/delete-service-category.controller";
 import { EditProductCategoryController } from "../../../Controllers/Products/Categories/update-service-category.controller";
 import { DeleteAllProductsCategoriesController } from "../../../Controllers/Products/Categories/delete-all-services-categories.controller";
+import { MiddlewareAuthorization } from "../../../Common/Middlewares/Authorization/authorization";
 
 const categoryRoutes: Router = Router()
 
 //Privates Services Categories Routes
-categoryRoutes.route("/products/categories/register-category").post((req: Request, res: Response) => {RegisterProductCategoryController.register(req, res)})
-categoryRoutes.route("/products/categories/edit-category/:id_category").patch((req: Request, res: Response) =>{EditProductCategoryController.editProduct(req, res)})
-categoryRoutes.route("/products/categories/delete-category/:id_category").delete((req: Request, res: Response) =>{DeleteProductCategoryController.delete(req, res)})
-categoryRoutes.route("/products/categories/delete-all-categories").delete((req: Request, res: Response) =>{DeleteAllProductsCategoriesController.deleteAll(req, res)})
+categoryRoutes.route("/products/categories/register-category").post(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isAdmin, (req: Request, res: Response) => {RegisterProductCategoryController.register(req, res)})
+categoryRoutes.route("/products/categories/edit-category/:id_category").patch(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isAdmin,(req: Request, res: Response) =>{EditProductCategoryController.editProduct(req, res)})
+categoryRoutes.route("/products/categories/delete-category/:id_category").delete(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isAdmin,(req: Request, res: Response) =>{DeleteProductCategoryController.delete(req, res)})
+categoryRoutes.route("/products/categories/delete-all-categories").delete(MiddlewareAuthorization.authorization, MiddlewareAuthorization.isAdmin,(req: Request, res: Response) =>{DeleteAllProductsCategoriesController.deleteAll(req, res)})
 
 //Publics Services Categories Routes
 categoryRoutes.route("/products/categories/get-all-categories").get((req: Request, res: Response) =>{GetProductsCategoriesController.findAll(req, res)})
